+++ src/sempipe/cli/embed_cmd.py
@@ -9,6 +9,7 @@ import sys
 import click
 
 from sempipe.cli.input_options import input_options, input_spec
+from sempipe.cli.interrupts import graceful_interrupts
 from sempipe.container import build_container
 from sempipe.core.errors import ExitCode
 from sempipe.verbs.embed import EmbedRequest, run_embed
@@ -47,5 +48,5 @@ def embed_command(
 
 
 async def _run(request: EmbedRequest) -> ExitCode:
-    async with build_container(os.environ) as container:
-        return await run_embed(request, container, stdin=sys.stdin, stdout=sys.stdout)
+    async with build_container(os.environ) as container, graceful_interrupts() as stop:
+        return await run_embed(request, container, stdin=sys.stdin, stdout=sys.stdout, stop=stop)